+++ wiswa/utils/__init__.py
@@ -2,7 +2,12 @@
 from __future__ import annotations
 
 from .github import setup_github_project
-from .jsonnet import evaluate_jsonnet_file, evaluate_jsonnet_project, evaluate_merged_settings
+from .jsonnet import (
+    evaluate_jsonnet_file,
+    evaluate_jsonnet_project,
+    evaluate_merged_settings,
+    resolve_defaults_only,
+)
 from .misc import create_py_typed_files
 from .path import non_empty_file_exists, primary_module_to_path, remove_empty_dirs
 from .postprocess import post_process_steps
@@ -13,5 +18,5 @@ from .versions import download_yarn, download_yarn_plugins, get_latest_yarn_vers
 __all__ = ('copy_static_files', 'create_py_typed_files', 'download_yarn', 'download_yarn_plugins',
            'evaluate_jsonnet_file', 'evaluate_jsonnet_project', 'evaluate_merged_settings',
            'get_latest_yarn_version', 'non_empty_file_exists', 'post_process_steps',
-           'primary_module_to_path', 'remove_empty_dirs', 'setup_github_project',
-           'write_templated_files')
+           'primary_module_to_path', 'remove_empty_dirs', 'resolve_defaults_only',
+           'setup_github_project', 'write_templated_files')